refactor: extract blog data fetching and metadata generation into dedicated library modules
